# server/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"JsonObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "upside-down",
		"version": "1.0.0",
		"description": "basic API using Node and Express",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/api/upsideDown/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/upsideDown/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"experience": {
										"type": "string"
									},
									"age": {
										"type": "string"
									},
									"character": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"experience",
									"age",
									"character",
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/upsideDown/update/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"experience": {
										"type": "string"
									},
									"age": {
										"type": "string"
									},
									"character": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"experience",
									"age",
									"character",
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/upsideDown/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/upsideDown/queryCharacter/{character}": {
			"get": {
				"operationId": "Query",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "character",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/upsideDown/getName/{name}": {
			"get": {
				"operationId": "QueryName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/upsideDown/getExperience/{experience}": {
			"get": {
				"operationId": "QueryExperience",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "experience",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/upsideDown/getAge/{age}": {
			"get": {
				"operationId": "QueryAge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "age",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}